move datasourse table into settings object
