+++ src/pudl/etl.py
@@ -13,7 +13,6 @@ data from:
  - US Environmental Protection Agency (EPA):
    - Continuous Emissions Monitory System (epacems)
 """
-import json
 import logging
 import time
 from concurrent.futures import ProcessPoolExecutor
@@ -446,76 +445,6 @@ def _read_static_pudl_tables() -> dict[str, pd.DataFrame]:
     return {"political_subdivisions": POLITICAL_SUBDIVISIONS}
 
 
-def make_datasources_table(datasets_settings, ds):
-    """Compile a table of dataset information.
-
-    There are three places we can look for information about a dataset:
-     * the datastore (for DOIs, working partitions, etc)
-     * the ETL settings (for partitions that are used in the ETL)
-     * the DataSource info (which is stored within the ETL settings)
-
-    The ETL settings and the datastore have different levels of nesting - and therefor
-    names for datasets. The nesting happens particularly with the EIA data. There
-    are three EIA datasets right now - eia923, eia860 and eia860m. eia860m is a monthly
-    update of a few tables in the larger eia860 dataset.
-    """
-    # grab all of the datasets that show up by name in the datastore
-    datasets_in_datastore_format = {
-        name: setting
-        for (name, setting) in datasets_settings.items()
-        if name in ds.get_known_datasets()
-    }
-    # add the eia datasets that are nested inside of the eia settings
-    if datasets_settings["eia"]:
-        datasets_in_datastore_format.update(
-            {
-                "eia923": datasets_settings["eia"].eia923,
-                "eia860": datasets_settings["eia"].eia860,
-            }
-        )
-
-    datasets = datasets_in_datastore_format.keys()
-    df = pd.DataFrame(
-        data={
-            "datasource": datasets,
-            "partitions": [
-                json.dumps(datasets_in_datastore_format[dataset].partitions)
-                for dataset in datasets
-            ],
-            "doi": [
-                _make_doi_clickable(ds.get_datapackage_descriptor(dataset).doi)
-                for dataset in datasets
-            ],
-        }
-    )
-    # add in EIA860m if eia in general is in the settings and the 860m bool is True
-    special_nested_datasets = pd.DataFrame()
-    if datasets_settings["eia"] and datasets_settings["eia"].eia860.eia860m:
-        special_nested_datasets = pd.DataFrame(
-            data={
-                "datasource": ["eia860m"],
-                "partitions": [
-                    json.dumps(
-                        datasets_in_datastore_format[
-                            "eia860"
-                        ].eia860m_data_source.working_partitions
-                    )
-                ],
-                "doi": [
-                    _make_doi_clickable(ds.get_datapackage_descriptor("eia860m").doi)
-                ],
-            }
-        )
-    df = pd.concat([df, special_nested_datasets]).reset_index(drop=True)
-    df["pudl_version"] = pudl.__version__
-    return df
-
-
-def _make_doi_clickable(link):
-    """Make a clickable DOI."""
-    return f"https://doi.org/{link}"
-
-
 ###############################################################################
 # Coordinating functions
 ###############################################################################
@@ -578,8 +507,8 @@ def etl(  # noqa: C901
         epacems_pq_path.mkdir(exist_ok=True)
 
     sqlite_dfs = _read_static_pudl_tables()
-    sqlite_dfs["datasources"] = make_datasources_table(
-        datasets_settings=datasets, ds=Datastore(**ds_kwargs)
+    sqlite_dfs["datasources"] = validated_etl_settings.make_datasources_table(
+        ds=Datastore(**ds_kwargs)
     )
     # This could be cleaner if we simplified the settings file format:
     if datasets.get("ferc1", False):
